removed na.rm during aggregation
I had earlier used na.rm when doing aggregation by day. This means that the day on which there was no observation became 0 instead of NA. This brought my median and mean down and it is not correct as no observation days should be simply ignored, not replaced by 0. (note: This is not to be confused with imput which is done later in the code.)